+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to **DecaLog** is documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- [MultiSite] Action link in sites list for network admins.
+- [MultiSite] Action link in "my sites" for local admins.
+
 ## [1.5.1] - 2019-10-05
 ### Changed
 - New logo, more in line with the plugin topic.